Moved image to assets, moved stylesheets to stylesheets
