+++ React/aethertex/src/pages/register.jsx
@@ -1,5 +1,5 @@
     import React from 'react';
-    import '../styling/register.css';
+    import '../stylesheets/register.css';
     import { Link } from 'react-router-dom';
 
     function Register() {
